+++ CHANGES.md
@@ -14,6 +14,7 @@ Deprecations:
 
 Language Improvements:
 
+- fix(javascript) fix regex inside parens after a non-regex (#2530) [Josh Goebel][]
 - [enh] Add `OPTIMIZE:` and `HACK:` to the labels highlighted inside comments [Josh Goebel][]
 - enh(typescript/javascript/coffeescript/livescript) derive ECMAscript keywords from a common foudation (#2518) [Josh Goebel][]
 - enh(typescript) add setInterval, setTimeout, clearInterval, clearTimeout (#2514) [Josh Goebel][]